Setup initial library functionality
Sets up the initial library functionality, which currently takes a
readable stream and returns one, doing the processing of the data in
rust (WASM) and deferring displaying data from the returned stream to
JS. This is NOT the most performant option and it is unclear at this
point what the boundary between JS and WASM will be. However, it does
start in the direction of acting on some concept of streams going to and
from devices, which may be via browser APIs or other sources (e.g. file
descriptors) in the future.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>